Timers added to all algorithms

+++ HelloWorld/algorithm.h
@@ -4,6 +4,7 @@
 #include <stack>
 #include "node.h"
 
+
 class Algorithm
 {
 private:
@@ -23,8 +24,8 @@ private:
 	void AddNode(Node** map1, Play::Vector2f parentID, int parentXplus, int parentYplus, float addValue, int corner, float parentValue, int width, int height, Play::Vector2f goal);
 
 public:
-	void aStar(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, int* runtime);
-	void Breadth(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, int* runtime);
-	void Depth(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, int* runtime);
-	void Dijkstra(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, int* runtime);
+	void aStar(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, long long* runtime);
+	void Breadth(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, long long* runtime);
+	void Depth(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, long long* runtime);
+	void Dijkstra(Node** map1, int width, int height, Play::Vector2f start, Play::Vector2f goal, long long* runtime);
 };
